Добавляет исправления сообщения при отправке фото
