--- js/data.js
@@ -1,69 +1,0 @@
-import {getRandomInteger, getRandomArrayElement} from './util.js';
-
-const AVATAR_ID_COUNT = 6;
-const LIKES_MIN_COUNT = 15;
-const LIKES_MAX_COUNT = 200;
-const COMMENTS_MAX_COUNT = 20;
-const OBJECT_COUNT = 25;
-
-const MESSAGES = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
-];
-
-const NAMES = [
-  'Иван',
-  'Хуан Себастьян',
-  'Мария',
-  'Кристоф',
-  'Виктор',
-  'Юлия',
-  'Люпита',
-  'Вашингтон',
-];
-
-const DESCRIPTIONS = [
-  'описание 1',
-  'описание 2',
-  'описание 3',
-  'описание 4',
-  'описание 5',
-  'описание 6',
-  'описание 7',
-];
-
-function createIdGenerator () {
-  let lastGeneratedId = 0;
-
-  return function () {
-    lastGeneratedId += 1;
-    return lastGeneratedId;
-  };
-}
-
-const generateId = createIdGenerator();
-const generatePhotosId = createIdGenerator();
-const generateCommentsId = createIdGenerator();
-
-const createComments = () => ({
-  id: generateCommentsId(),
-  avatar: `img/avatar-${ getRandomInteger(1, AVATAR_ID_COUNT) }.svg`,
-  message: getRandomArrayElement(MESSAGES),
-  name: getRandomArrayElement(NAMES),
-});
-
-const createObject = () => ({
-  id: generateId(),
-  url: `photos/${generatePhotosId()}.jpg`,
-  description: getRandomArrayElement(DESCRIPTIONS),
-  likes: getRandomInteger(LIKES_MIN_COUNT, LIKES_MAX_COUNT),
-  comments: Array.from({length: getRandomInteger(1, COMMENTS_MAX_COUNT)}, createComments),
-});
-
-const createObjects = (count) => Array.from({length: count}, createObject);
-
-export {createObjects, OBJECT_COUNT};
